fix: update man files, just
This commit updates manual file templates, as well as the man-page build
script using just to reflect the interim project name.

This is necessary to provide manual files that are accurate.

It should be somewhat easy to revert revert these changes, and is
necessary to document new features as they are merged.

Signed-off-by: Christina Sørensen <christina@cafkafk.com>